Ajuste de entrada e retorno dos metodos service para receber dtos e nao a entidade/model.

+++ padaria/padaria/src/main/java/com/louise/padaria/service/ProdutoService.java
@@ -1,10 +1,16 @@
 package com.louise.padaria.service;
 
+import com.louise.padaria.dto.ProdutoConsultarDto;
+import com.louise.padaria.dto.ProdutoDto;
+import com.louise.padaria.dto.ProdutoEditarDto;
+import com.louise.padaria.dto.ProdutoSalvarDto;
+import com.louise.padaria.mapper.ProdutoMapper;
 import com.louise.padaria.model.Produto;
 import com.louise.padaria.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,40 +18,53 @@ import java.util.Optional;
 public class ProdutoService {
     @Autowired
     private ProdutoRepository repository;
+    @Autowired
+    private ProdutoMapper mapper;
 
-    public boolean salvarProduto(Produto produto){
+    public boolean salvarProduto(ProdutoSalvarDto dto){
+        Produto produto = mapper.converteDtoParaModel(dto);
         repository.save(produto);
         return true;
     }
 
     public boolean deletarProduto(Integer pr){
-        Produto produto = buscarProdutoPorId(pr);
-        if(produto != null){
-            repository.delete(produto);
-            return true;
+        if(pr != null) {
+            Produto produto = mapper.converteDtoParaModel(buscarProdutoPorId(pr));
+            if (produto != null) {
+                repository.delete(produto);
+                return true;
+            }
         }
         return false;
     }
-    public boolean atualizarProduto(Produto prod){
-        Produto produto = buscarProdutoPorId(prod.getId());
-        if (produto != null){
-            repository.save(prod);
-            return true;
+    public boolean atualizarProduto(ProdutoEditarDto prod){
+        if(prod != null){
+            ProdutoConsultarDto produto = buscarProdutoPorId(prod.getId());
+            if (produto != null){
+                 Produto produtoAtualizar = mapper.converteDtoParaModel(prod);
+                 repository.save(produtoAtualizar;
+                 return true;
+            }
         }
         return false;
     }
 
-    public Produto buscarProdutoPorId(Integer id){
+    public ProdutoConsultarDto buscarProdutoPorId(Integer id){
+        if(id == null){
+            return null;
+        }
        Optional<Produto>produto = repository.findById(id);
        if(produto.isPresent()){
-           return produto.get();
+           ProdutoConsultarDto dto = mapper.converteModelParaConsultar(produto.get());
+           return dto;
        }
         return null;
-
     }
-    public List<Produto> buscarTodosProdutos(){
-
-        return repository.findAll();
+    public List<ProdutoDto> buscarTodosProdutos(){
+        List<ProdutoDto>lista = new ArrayList<>();
+        for (Produto p: repository.findAll()){
+            lista.add(mapper.converteModelParaDto(p));
+        }
+        return lista;
     }
-
 }
